requirements.txt
adafruit-board-toolkit==1.1.1
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
astroid==2.15.8
asttokens==2.2.1
bcrypt==3.2.2
beautifulsoup4==4.14.3
bitarray==2.8.3
bitstring==4.1.2
blinker==1.9.0
certifi==2026.2.25
cffi==1.16.0
charset-normalizer==3.4.6
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
cryptography==38.0.4
dask==2024.8.2
dill==0.3.7
distro==1.9.0
docstring_parser==0.17.0
docutils==0.20.1
ecdsa==0.18.0
esptool==4.6.2
exceptiongroup==1.3.1
Flask==3.1.3
fsspec==2024.6.1
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.69.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
imageio==2.35.1
importlib_metadata==8.4.0
impy-array==2.4.4
isort==5.12.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lazy-object-proxy==1.9.0
lazy_loader==0.4
locket==1.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.4
numpy==2.1.0
openai==2.30.0
packaging==24.1
pandas==2.2.2
paramiko==3.2.0
parso==0.8.3
partd==1.4.2
pickle4==0.0.1
pillow==10.4.0
platformdirs==4.0.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
pylint==2.17.7
PyNaCl==1.5.0
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2024.1
PyYAML==6.0.1
reedsolo==1.7.0
regex==2026.3.32
requests==2.33.1
scikit-image==0.24.0
scikit-learn==1.7.2
scipy==1.14.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
thonny==4.1.4
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.7
wrapt==1.16.0
zipp==3.20.1
